refactor: extract /character and /jobs data assembly into shared seams
Phase 1 of docs/registrations-page: move the data assembly out of the two
page.tsx files so /registration can import the same objects rather than a
copy that would drift.

- src/app/character/characterData.ts — fetchCharacterOverviews() returns one
  CharacterOverview per registration (ISK, location, ship, clones, implants,
  job slots), plus hasNoOptionalScopes() for the limited-access warning.
- src/app/character/jobSlotBubbles.tsx — the JobSlots bubbles, split out so
  the new page renders the identical component (named to not read as a second
  src/app/industry/jobSlots.ts).
- src/app/jobs/jobsData.ts — fetchJobsOverview() with the heartbeat/open-run/
  refresh_task fold, corp grouping and representative picking, runFor().

Both pages keep their auth gate and all JSX. Pure refactor: no rendered byte
changes.
